backend/requirements.txt
asgiref==3.11.0
Django==5.2.1
django-allauth==0.63.6
django-cors-headers==4.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dj-rest-auth==6.0.0
pillow==10.3.0
sqlparse==0.5.0
tzdata==2024.1
